+++ challangeNetflix/ ViewController.swift
@@ -24,40 +24,46 @@ class ViewController: UIViewController, UICollectionViewDelegate, UICollectionVi
     let session = URLSession.shared
     let url = URL(string: "http://localhost:8080/response.json")!
     var movieSelected: Movie?
+    let movieAux: MovieController = MovieController()
     
     var searchView = UICollectionReusableView()
     
     override func viewDidLoad() {
         super.viewDidLoad()
 		
-        let task = session.dataTask(with: url) {
-            data, response, error in
+//        let task = session.dataTask(with: url) {
+//            data, response, error in
+//
+//            guard let httpResponse = response as? HTTPURLResponse,
+//                (200...299).contains(httpResponse.statusCode) else {
+//                    print(error)
+//                    return
+//            }
+//        }
+        
+        //executar a task atraves da func request??? aqui preciso da data
+        
+        var data = movieAux.request(urlName: "http://localhost:8080/response.json")
+        
+        do {
+            self.movies = try self.decoder.decode([Movie].self, from: data)
             
-            guard let httpResponse = response as? HTTPURLResponse,
-                (200...299).contains(httpResponse.statusCode) else {
-                    print(error)
-                    return
+            var database: MovieDatabase = MovieDatabase(config: Realm.Configuration())
+            for movie in self.movies {
+                database.addFilmDB(movie: movie)
             }
             
-            do {
-                self.movies = try self.decoder.decode([Movie].self, from: data!)
-                
-                var database: MovieDatabase = MovieDatabase(config: Realm.Configuration())
-                for movie in self.movies {
-                    database.addFilmDB(movie: movie)
-                }
-                
-                DispatchQueue.main.async {
-                    self.realData = self.movies
-                    self.collectionView.reloadData()
-                }
-            } catch {
-                print(error)
+            DispatchQueue.main.async {
+                self.realData = self.movies
+                self.collectionView.reloadData()
             }
+        } catch {
+            print(error)
         }
+        
         collectionView.delegate = self
         collectionView.dataSource = self
-        task.resume()
+        
         
     }
     
@@ -120,25 +126,7 @@ class ViewController: UIViewController, UICollectionViewDelegate, UICollectionVi
         }
     }
     
-    //    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-    //        self.movies.removeAll()
-    //
-    //        for item in self.realData {
-    //            if (item.title.lowercased().contains(searchBar.text!.lowercased())) {
-    //                self.movies.append(item)
-    //            }
-    //        }
-    //
-    //        if (searchBar.text!.isEmpty) {
-    //            self.movies = self.realData
-    //        }
-    //
-    //        self.collectionView.reloadData()
-    //    }
-    
     func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-        
-        let movieAux: MovieAux = MovieAux()
         
         self.movies = movieAux.searchMovie(movies: realData, search: searchBar.text!)
