+++ challangeNetflix/MovieDatabase.swift
@@ -50,17 +50,17 @@ class MovieDatabase {
         
         var allObjects: [MovieRealm] = []
         
-        do {
+//        do {
             
             var realmResults = realm.objects(MovieRealm.self)
             
             for movieRealm in realmResults {
                 allObjects.append(movieRealm)
             }
-            
-        } catch let error as NSError {
-            print(error)
-        }
+//
+//        } catch let error as NSError {
+//            print(error)
+//        }
         
         return allObjects
     }
